shui/requirements.txt
python-fasthtml
pillow
requests